implement the YjsService
