+++ websocket/app/src/lib.rs
@@ -8,3 +8,4 @@ mod routes;
 pub use routes::create_router;
 mod middleware;
 pub use middleware::add_middleware;
+mod services;
